Split languages.py into per-language package
Reorganize src/literalizer/languages.py into a languages/ package with one file per language. This reduces the largest file from 1242 lines to individual focused files. Each language module imports only the formatters and dependencies it needs, improving maintainability and reducing cognitive load when modifying language specifications.
