+++ src/literalizer/languages/__init__.py
@@ -1,0 +1,61 @@
+"""Built-in language specifications for common programming languages."""
+
+from __future__ import annotations
+
+from .c import C
+from .clojure import CLOJURE
+from .cpp import CPP
+from .csharp import CSHARP
+from .dart import DART
+from .elixir import ELIXIR
+from .erlang import ERLANG
+from .fsharp import FSHARP
+from .go import GO
+from .groovy import GROOVY
+from .haskell import HASKELL
+from .java import JAVA
+from .javascript import JAVASCRIPT
+from .julia import JULIA
+from .kotlin import KOTLIN
+from .lua import LUA
+from .ocaml import OCAML
+from .occam import OCCAM
+from .perl import PERL
+from .php import PHP
+from .python import PYTHON
+from .r import R
+from .ruby import RUBY
+from .rust import RUST
+from .scala import SCALA
+from .swift import SWIFT
+from .typescript import TYPESCRIPT
+
+__all__ = [
+    "CLOJURE",
+    "CPP",
+    "CSHARP",
+    "DART",
+    "ELIXIR",
+    "ERLANG",
+    "FSHARP",
+    "GO",
+    "GROOVY",
+    "HASKELL",
+    "JAVA",
+    "JAVASCRIPT",
+    "JULIA",
+    "KOTLIN",
+    "LUA",
+    "OCAML",
+    "OCCAM",
+    "PERL",
+    "PHP",
+    "PYTHON",
+    "RUBY",
+    "RUST",
+    "SCALA",
+    "SWIFT",
+    "TYPESCRIPT",
+    "C",
+    "R",
+]
