+++ src/literalizer/languages/csharp.py
@@ -1,0 +1,56 @@
+"""C# language specification."""
+
+from __future__ import annotations
+
+from beartype import beartype
+
+from literalizer._language import Language
+from literalizer.formatters import (
+    format_bytes_hex,
+    format_date_iso,
+    format_datetime_iso,
+    format_variable_assignment_csharp,
+    format_variable_declaration_csharp,
+    passthrough_sequence_entry,
+    passthrough_set_entry,
+)
+
+
+@beartype
+def _format_csharp_dict_entry(key: str, value: str) -> str:
+    """Format a C# dictionary indexer entry."""
+    return f"[{key}] = {value}"
+
+
+CSHARP = Language(
+    null_literal="(object?)null",
+    true_literal="true",
+    false_literal="false",
+    sequence_open="(",
+    sequence_close=")",
+    dict_open="new Dictionary<string, object> {",
+    dict_close="}",
+    format_dict_entry=_format_csharp_dict_entry,
+    multiline_trailing_comma=False,
+    single_element_trailing_comma=False,
+    format_bytes=format_bytes_hex,
+    format_date=format_date_iso,
+    format_datetime=format_datetime_iso,
+    empty_sequence="ValueTuple.Create()",
+    empty_dict=None,
+    set_open="new HashSet<object> {",
+    set_close="}",
+    empty_set="new HashSet<object>()",
+    format_sequence_entry=passthrough_sequence_entry,
+    format_set_entry=passthrough_set_entry,
+    comment_prefix="//",
+    comment_suffix="",
+    omap_open="new Dictionary<string, object> {",
+    omap_close="}",
+    format_omap_entry=_format_csharp_dict_entry,
+    multiline_close_indent="",
+    element_separator=", ",
+    skip_null_dict_values=False,
+    format_variable_declaration=format_variable_declaration_csharp,
+    format_variable_assignment=format_variable_assignment_csharp,
+)
