+++ src/literalizer/languages/java.py
@@ -1,0 +1,56 @@
+"""Java language specification."""
+
+from __future__ import annotations
+
+from beartype import beartype
+
+from literalizer._language import Language
+from literalizer.formatters import (
+    format_bytes_hex,
+    format_date_iso,
+    format_datetime_iso,
+    format_variable_assignment_java,
+    format_variable_declaration_java,
+    passthrough_sequence_entry,
+    passthrough_set_entry,
+)
+
+
+@beartype
+def _format_java_dict_entry(key: str, value: str) -> str:
+    """Format a Java ``Map.entry(key, value)`` call."""
+    return f"Map.entry({key}, {value})"
+
+
+JAVA = Language(
+    null_literal="null",
+    true_literal="true",
+    false_literal="false",
+    sequence_open="new Object[]{",
+    sequence_close="}",
+    dict_open="Map.ofEntries(",
+    dict_close=")",
+    format_dict_entry=_format_java_dict_entry,
+    multiline_trailing_comma=False,
+    single_element_trailing_comma=False,
+    format_bytes=format_bytes_hex,
+    format_date=format_date_iso,
+    format_datetime=format_datetime_iso,
+    empty_sequence=None,
+    empty_dict=None,
+    set_open="Set.of(",
+    set_close=")",
+    empty_set=None,
+    format_sequence_entry=passthrough_sequence_entry,
+    format_set_entry=passthrough_set_entry,
+    comment_prefix="//",
+    comment_suffix="",
+    omap_open="new java.util.ArrayList<>(java.util.Arrays.asList(",
+    omap_close="))",
+    format_omap_entry=_format_java_dict_entry,
+    multiline_close_indent="",
+    element_separator=", ",
+    format_variable_declaration=format_variable_declaration_java,
+    format_variable_assignment=format_variable_assignment_java,
+    skip_null_dict_values=True,
+)
