+++ src/literalizer/languages/python.py
@@ -1,0 +1,54 @@
+"""Python language specification."""
+
+from __future__ import annotations
+
+from literalizer._language import Language
+from literalizer.formatters import (
+    dict_entry_with_separator,
+    format_bytes_hex,
+    format_date_iso,
+    format_datetime_iso,
+    format_variable_assignment_python,
+    format_variable_declaration_python,
+    passthrough_sequence_entry,
+    passthrough_set_entry,
+)
+
+
+def _format_python_omap_entry(key: str, value: str) -> str:
+    """Format one Python ``OrderedDict`` entry as a ``(key, value)`` tuple."""
+    return f"({key}, {value})"
+
+
+PYTHON = Language(
+    null_literal="None",
+    true_literal="True",
+    false_literal="False",
+    sequence_open="(",
+    sequence_close=")",
+    dict_open="{",
+    dict_close="}",
+    format_dict_entry=dict_entry_with_separator(separator=": "),
+    multiline_trailing_comma=True,
+    single_element_trailing_comma=True,
+    format_bytes=format_bytes_hex,
+    format_date=format_date_iso,
+    format_datetime=format_datetime_iso,
+    empty_sequence=None,
+    empty_dict=None,
+    set_open="{",
+    set_close="}",
+    empty_set="set()",
+    format_sequence_entry=passthrough_sequence_entry,
+    format_set_entry=passthrough_set_entry,
+    comment_prefix="#",
+    comment_suffix="",
+    omap_open="OrderedDict([",
+    omap_close="])",
+    format_omap_entry=_format_python_omap_entry,
+    multiline_close_indent="",
+    element_separator=", ",
+    skip_null_dict_values=False,
+    format_variable_declaration=format_variable_declaration_python,
+    format_variable_assignment=format_variable_assignment_python,
+)
