+++ src/literalizer/languages/rust.py
@@ -1,0 +1,61 @@
+"""Rust language specification."""
+
+from __future__ import annotations
+
+from beartype import beartype
+
+from literalizer._language import Language
+from literalizer.formatters import (
+    format_bytes_hex,
+    format_date_iso,
+    format_datetime_iso,
+    format_variable_assignment_rust,
+    format_variable_declaration_rust,
+    passthrough_sequence_entry,
+    passthrough_set_entry,
+)
+
+
+@beartype
+def _format_rust_dict_entry(key: str, value: str) -> str:
+    """Format a Rust HashMap entry as a tuple ``(key, value)``."""
+    return f"({key}, {value})"
+
+
+def _format_rust_omap_entry(key: str, value: str) -> str:
+    """Format a Rust ordered-map entry as a tuple ``(key, value)``."""
+    return f"({key}, {value})"
+
+
+RUST = Language(
+    null_literal="None",
+    true_literal="true",
+    false_literal="false",
+    sequence_open="vec![",
+    sequence_close="]",
+    dict_open="HashMap::from([",
+    dict_close="])",
+    format_dict_entry=_format_rust_dict_entry,
+    multiline_trailing_comma=True,
+    single_element_trailing_comma=False,
+    format_bytes=format_bytes_hex,
+    format_date=format_date_iso,
+    format_datetime=format_datetime_iso,
+    empty_sequence=None,
+    empty_dict=None,
+    set_open="HashSet::from([",
+    set_close="])",
+    empty_set=None,
+    format_sequence_entry=passthrough_sequence_entry,
+    format_set_entry=passthrough_set_entry,
+    comment_prefix="//",
+    comment_suffix="",
+    omap_open="HashMap::from([",
+    omap_close="])",
+    format_omap_entry=_format_rust_omap_entry,
+    multiline_close_indent="",
+    element_separator=", ",
+    skip_null_dict_values=False,
+    format_variable_declaration=format_variable_declaration_rust,
+    format_variable_assignment=format_variable_assignment_rust,
+)
